requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via litellm
aiosignal==1.3.2
    # via aiohttp
alembic==1.14.0
    # via yayska (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
asyncpg==0.30.0
    # via yayska (pyproject.toml)
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
bcrypt==4.2.1
    # via passlib
cachetools==5.5.0
    # via
    #   yayska (pyproject.toml)
    #   google-auth
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   litellm
    #   uvicorn
cryptography==44.0.2
    # via python-jose
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.2.0
    # via pydantic
fastapi==0.115.6
    # via
    #   yayska (pyproject.toml)
    #   fastapi-cache2
fastapi-cache2==0.2.2
    # via yayska (pyproject.toml)
filelock==3.18.0
    # via huggingface-hub
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.5.1
    # via huggingface-hub
google-auth==2.38.0
    # via google-auth-oauthlib
google-auth-oauthlib==1.2.1
    # via yayska (pyproject.toml)
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.1.5
    # via huggingface-hub
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   yayska (pyproject.toml)
    #   litellm
    #   openai
huggingface-hub==0.33.1
    # via tokenizers
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via litellm
jinja2==3.1.6
    # via litellm
jiter==0.10.0
    # via openai
jsonschema==4.24.0
    # via litellm
jsonschema-specifications==2025.4.1
    # via jsonschema
litellm==1.73.2
    # via yayska (pyproject.toml)
mako==1.3.8
    # via alembic
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
multidict==6.6.0
    # via
    #   aiohttp
    #   yarl
oauthlib==3.2.2
    # via requests-oauthlib
openai==1.93.0
    # via litellm
packaging==25.0
    # via huggingface-hub
passlib==1.7.4
    # via yayska (pyproject.toml)
pendulum==3.0.0
    # via fastapi-cache2
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
psycopg2-binary==2.9.10
    # via yayska (pyproject.toml)
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   python-jose
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.10.4
    # via
    #   yayska (pyproject.toml)
    #   fastapi
    #   litellm
    #   openai
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.1
    # via yayska (pyproject.toml)
pyjwt==2.10.1
    # via yayska (pyproject.toml)
python-dateutil==2.9.0.post0
    # via
    #   pendulum
    #   time-machine
python-dotenv==1.0.1
    # via
    #   yayska (pyproject.toml)
    #   litellm
    #   pydantic-settings
    #   uvicorn
python-jose==3.4.0
    # via yayska (pyproject.toml)
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   uvicorn
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   huggingface-hub
    #   requests-oauthlib
    #   tiktoken
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via
    #   google-auth
    #   python-jose
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.36
    # via
    #   yayska (pyproject.toml)
    #   alembic
starlette==0.41.3
    # via fastapi
structlog==24.4.0
    # via yayska (pyproject.toml)
tenacity==8.5.0
    # via yayska (pyproject.toml)
tiktoken==0.9.0
    # via litellm
time-machine==2.16.0
    # via pendulum
tokenizers==0.21.2
    # via litellm
tqdm==4.67.1
    # via
    #   yayska (pyproject.toml)
    #   huggingface-hub
    #   openai
typing-extensions==4.12.2
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   fastapi-cache2
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
tzdata==2024.2
    # via pendulum
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via
    #   yayska (pyproject.toml)
    #   fastapi-cache2
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.3
    # via uvicorn
websockets==14.1
    # via uvicorn
yarl==1.20.1
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
